Added support for paired fastq files.
Paired fastQ files can now be uploaded and overlap-merged.

The algorithm for overlap merge is as follows:
1. Build a kmer to position dict of read 2
2. Identify HSP regions in between the two reads of a given length
3. Splice the two reads together at the longest HSP.

HSP search and identification is also controlled by requirements on minimum number of
adjacent kmers to form an HSP, kmer length and minimum number of kmers, being part of HSPs
that make up the overlap.