Refactoring of primer removal code.
The primer removal code has ben broken out to the FilterSeq class and optimised.

Primers are now searched in a much more efficient way by encoding the bases as one base per bit in lists. This also
enabled matching of ambiguite bases in the primer sequence by simple bitwise AND operations.

Two other constraints added are that the 5' primer must be found within the first 100 bp and that the
primer is considered to be found at the first sequence match that is good enough.

The speedup of the new code is roughly 60 times.